Experiment with a Cow-ified version of uaparser
I've found that the uaparser we use for "extended" parsing of user agents makes
a number of small allocations internally. I've got an experimental branch of
that project that prefers the use of `Cow` instead; this commit targets that
branch. I can measure a micro improvement in the benchmarks of uaparser but it's
unclear if this change will positively impact vector. Notably, even though
uaparser is returning a `Cow` this library then causes the small string to go
through an allocation, since our types are `String` on the VRL side.

REF #12267

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>