Use `value::Value` directly in the codebase
This commit modifies the project to use `value::Value` directly and not as a
re-export from VRL. The changeset is large but almost entirely mechanical. I
would like to experiment with different representations of `Value` -- imagine a
world where the `Value` is mostly stack allocated, or where the `Value::Object`
doesn't recurse to another heap allocated `Value` outside of the current virtual
machine page -- but I realized while making some modifications to that type that
its representation was, well, kindly spread through the project.

We don't solve that here. Instead we take a single step in that
direction. Follow-up work will work a struct shim over the top of the `BTreeMap`
inside `Value` so callers are not aware of the `BTreeMap`. That abstraction is
what will allow us to fiddle with the representation, so long as we hew to the
enum nature of `Value`.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>